medical_search/requirements.txt
streamlit==1.23.1
streamlit-aggrid==0.3.4.post3
pandas==2.0.2
google-cloud-discoveryengine==0.9.0
google-cloud-aiplatform==1.26.0
google-cloud-storage==2.9.0